# documentacion/sgNetApi_OpenAPI_v3.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "sgNetApi - Módulo de Administración y Seguridad",
    "description": "Especificación OpenAPI v3 para el Módulo de Administración, Seguridad y Auditoría del sistema sgNetApi.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5283",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "paths": {
...
    "/api/Usuarios": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Obtener lista completa de usuarios",
        "description": "Retorna el catálogo general de usuarios con sus grados, escalafones y roles asociados.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios retornada con éxito"
          },
          "401": {
            "description": "No autorizado"
          },
          "403": {
            "description": "Permiso 'admin.usuarios.leer' requerido"
          }
        }
      },
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Crear un nuevo usuario",
        "description": "Registra un nuevo funcionario en la base de datos.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ci": {
                    "type": "integer",
                    "format": "int64",
                    "example": 12345678
                  },
                  "nombre": {
                    "type": "string",
                    "example": "Juan"
                  },
                  "apellido": {
                    "type": "string",
                    "example": "Pérez"
                  },
                  "correo": {
                    "type": "string",
                    "example": "juan.perez@sgnet.gub.uy"
                  },
                  "celular": {
                    "type": "integer",
                    "example": 99111222
                  },
                  "idGrado": {
                    "type": "integer",
                    "example": 1
                  },
                  "idEscalafon": {
                    "type": "integer",
                    "example": 1
                  },
                  "idUuee": {
                    "type": "integer",
                    "example": 1
                  },
                  "idDependencia": {
                    "type": "integer",
                    "example": 1
                  },
                  "idsRoles": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      2
                    ]
                  },
                  "idsPermisosDirectos": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": []
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente"
          },
          "403": {
            "description": "Permiso 'admin.usuarios.crear' requerido"
          }
        }
      }
    },
    "/api/Usuarios/{ci}/roles-permisos": {
      "put": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Actualizar roles y permisos de un usuario",
        "description": "Sincroniza atómicamente los roles y permisos directos asignados a un usuario específico.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ci",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idsRoles": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      1,
                      2
                    ]
                  },
                  "idsPermisosDirectos": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    },
                    "example": [
                      6
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Roles y permisos actualizados correctamente"
          },
          "403": {
            "description": "Permiso 'admin.usuarios.editar' requerido"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}